requirements.txt
discord==2.3.2
discord.py==2.4.0
aiohttp==3.9.5
python-dotenv==1.0.1
pymongo==4.7.3
motor==3.4.0
pydantic==2.7.4
transformers
torch
torchaudio
torchvision
pillow==10.2.0
numpy==1.26.4
opencv-python==4.10.0.84
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
bitsandbytes==0.43.1
diffusers==0.29.1
huggingface-hub==0.23.4
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.13.1
tqdm==4.66.2
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.22.0
decord
pdf2image
opencc
google-generativeai
langchain_huggingface
sentencepiece==0.2.0
vector-quantize-pytorch==1.18.5
vocos==0.1.0
accelerate==1.2.1
timm==0.9.10
soundfile==0.12.1
librosa==0.9.0
decord
moviepy
webdriver-manager